README.md
Library_management
Library_management is a software used to manages the catalog of a library. This helps to keep the records of whole transactions of the books available in the library.

AmpleTrails provides Library Management System which is very easy to use and fulfills all the requirement of a librarian. There are many features which helps librarian to keep records of available books as well as issued books. This software is available in both mode i.e. web-based or local host based. We provide best Library Management System of this planet.

Here is a list of some features of Library Management System which AmpleTrails offer:

1.Keep record of different categories like; Books, Journals, Newspapers, Magazines, etc.

2.Classify the books subject wise.

3.Easy way to enter new books.

4.Keep record of complete information of a book like; Book name, Author name, Publisher’s name, Date/ Year of publication, Cost of the book, Book purchasing date/ Bill no.

5.Easy way to make a check-out.

6.Easy way to make a check-in.

7.Automatic fine calculation for late returns.

8.Different criteria for searching a book.

9.Different kind of reports like; total no. of books, no. of issued books, no. of journals, etc.

10.Easy way to know how many books are issued to a particular student.

and much more.